report order done

+++ application/views/admin/order/index.php
@@ -16,8 +16,8 @@
         <?php foreach($orders as $order): ?>
           <tr>
             <td><?php echo $order->id; ?></td>
-            <td><?php echo $order->id_room; ?></td>
-            <td><?php echo $order->date; ?></td>
+            <td><?php echo $order->idRoom; ?></td>
+            <td><?php echo date("d-m-Y",strtotime($order->date)); ?></td>
             <td><?php echo $order->total; ?></td>
             <td>
             <a href="<?= base_url(); ?>admin/order/detail/<?= $order->id ;?>" class="btn btn-primary">
@@ -26,7 +26,7 @@
               <a href="<?= base_url(); ?>admin/order/edit/<?= $order->id ;?>" class="btn btn-warning">
                 <i class="fa fa-pen" aria-hidden="true"></i>
               </a>
-              <a href="<?= base_url(); ?>admin/order/hapus/<?= $order->id ;?>" class="btn btn-danger">
+              <a href="<?= base_url(); ?>admin/order/hapus/<?= $order->id ;?>" class="btn btn-danger" onClick="return confirm('Apakah Kamu yakin Akan Menghapus ??');">
                 <i class="fa fa-trash" aria-hidden="true"></i>
               </a>
             </td>
